+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 - nostr: update `Nip19Profile::new` and `Nip19Coordinate::new` signature (https://github.com/rust-nostr/nostr/pull/910)
 - nostr: update `RelayInformationDocument::get` signature (https://github.com/rust-nostr/nostr/pull/913)
 - nostr: remove `TagStandard::Delegation` and `TagKind::Delegation` (https://github.com/rust-nostr/nostr/pull/929)
+- nostr: remove `nip05` cargo feature (https://github.com/rust-nostr/nostr/pull/936)
+- nostr: convert `nip05` module to be I/O-free (https://github.com/rust-nostr/nostr/pull/936)
 - connect: remove `NostrConnect::get_relays` (https://github.com/rust-nostr/nostr/pull/894)
 - database: merge traits into `NostrDatabase` (https://github.com/rust-nostr/nostr/pull/916)
 - database: remove `NostrDatabase::has_coordinate_been_deleted` (https://github.com/rust-nostr/nostr/pull/917)